requirements.txt
# UI Testing
selenium==4.15.0
webdriver-manager==4.0.1
Pillow==10.1.0

# API Testing
requests==2.31.0
jsonschema==4.20.0
pydantic==2.5.0

# Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-retry==1.6.1
allure-pytest==2.13.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1

# Additional Testing Tools
pytest-mock==3.12.0
pytest-cov==4.1.0
faker==20.1.0

# Logging and Reporting
colorlog==6.8.0


